Start UpdateTour (error with image replace), Display time and distance

+++ TourPlanner/TourPlanner.DataAccessLayer.REST/HttpRequest.cs
@@ -24,7 +24,7 @@ namespace TourPlanner.DataAccessLayer.REST
 
 		    var json = JsonNode.Parse(await _client.GetStringAsync(url));
 		    tour.Distance = json["route"]["distance"].GetValue<double>();
-		    tour.EstimatedTime = json["route"]["time"].GetValue<double>();
+		    tour.EstimatedTime = json["route"]["time"].GetValue<int>();
 
 		    return tour;
 		}
